Track dependency conditions when parsing cnditional depends_on
This commit changes the way that depends_on blocks are parsed and
tracked during the `podman-compose up` process.

Each service's dependencies will be tracked as a dict in which keys are
services being depended on and values are the condition to be met by
said services before the current service can start.

This lays the groundwork for supporting long-syntax / conditional
depends_on blocks, but should not change any behavior so far.

Signed-off-by: Adrian Torres <atorresj@redhat.com>